+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Support for ',' delimiter for loading the data.
 - Support for loading the data from a 'ng_trajectory' csv file.
 - Callback for `input_file` so that the file can be changed during runtime.
+- Subscriber for `move_base_simple/goal` to rotate and translate the trajectory.
 
 ### Changed
 - Node is now executable using Python 3 as well.